+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 1.0.0 
+## 1.1.0
+
+### New features
+- New property **startPosition** for **ControlledAnimation** widget:
+  you can now specify the initial start position (time) of the animation.
+  This is useful for designing interactive UI elements.
+
+
+## 1.0.0
 - Initial release
 
 ### New features